#45  Move the rest of create table test to CreateTableParserTest
