+++ cql-parser/src/test/scala/troy/cql/parser/CqlParserTest.scala
@@ -22,83 +22,7 @@ import troy.cql.ast.ddl.{ Keyspace, KeyspaceName, Table }
 
 class CqlParserTest extends FlatSpec with Matchers {
 
-  "Cql Parser" should "parse multiline create table" in {
-    val statement = parseSchemaAs[CreateTable](
-      """
-        CREATE TABLE test.posts (
-          author_id text PRIMARY KEY
-        );
-      """
-    )
-    statement.ifNotExists shouldBe false
-    statement.columns.size shouldBe 1
-    statement.columns.head.dataType shouldBe DataType.text
-    statement.columns.head.isPrimaryKey shouldBe true
-    statement.columns.head.isStatic shouldBe false
-    statement.columns.head.name shouldBe "author_id"
-  }
-
-  it should "parse multi-fields create table" in {
-    val statement = parseSchemaAs[CreateTable](
-      """
-        CREATE TABLE test.posts (
-          author_id text PRIMARY KEY,
-          author_name text static,
-          author_age int static,
-          post_id text,
-          post_title text
-        );
-      """
-    )
-    statement.ifNotExists shouldBe false
-    statement.columns shouldBe Seq(
-      Table.Column("author_id", DataType.text, false, true),
-      Table.Column("author_name", DataType.text, true, false),
-      Table.Column("author_age", DataType.int, true, false),
-      Table.Column("post_id", DataType.text, false, false),
-      Table.Column("post_title", DataType.text, false, false)
-    )
-    statement.primaryKey.isEmpty shouldBe true // Primary is defined inline instead
-    statement.options.isEmpty shouldBe true
-  }
-
-  it should "parse create table with clustering columns" in {
-    val statement = parseSchemaAs[CreateTable](
-      """
-        CREATE TABLE test.posts (
-          author_id text,
-          author_name text static,
-          author_age int static,
-          post_id text,
-          post_title text,
-          PRIMARY KEY ((author_id), post_id)
-        );
-      """
-    )
-    statement.ifNotExists shouldBe false
-    statement.columns shouldBe Seq(
-      Table.Column("author_id", DataType.text, false, false), // Yes, it is not THE primary key, only a partition key
-      Table.Column("author_name", DataType.text, true, false),
-      Table.Column("author_age", DataType.int, true, false),
-      Table.Column("post_id", DataType.text, false, false),
-      Table.Column("post_title", DataType.text, false, false)
-    )
-    statement.primaryKey.get.partitionKeys shouldBe Seq("author_id")
-    statement.primaryKey.get.clusteringColumns shouldBe Seq("post_id")
-  }
-
-  it should "parse different primary key styles" in {
-    def pk(pk: String) =
-      parseSchemaAs[CreateTable](s"CREATE TABLE test.posts ( author_id text, author_name text, post_id text, post_title text, PRIMARY KEY $pk);").primaryKey.get
-
-    pk("((author_id), post_id)") shouldBe Table.PrimaryKey(Seq("author_id"), Seq("post_id"))
-    pk("(author_id, post_id)") shouldBe Table.PrimaryKey(Seq("author_id"), Seq("post_id"))
-    pk("((author_id, author_name), post_id)") shouldBe Table.PrimaryKey(Seq("author_id", "author_name"), Seq("post_id"))
-    pk("((author_id, author_name), post_id, post_title)") shouldBe Table.PrimaryKey(Seq("author_id", "author_name"), Seq("post_id", "post_title"))
-    pk("(author_id, post_id, post_title)") shouldBe Table.PrimaryKey(Seq("author_id"), Seq("post_id", "post_title"))
-  }
-
-  it should "parse multiple statements " in {
+  "Cql Parser" should "parse multiple statements " in {
     val statements = parseSchema(
       """
             CREATE KEYSPACE test WITH replication = {'class': 'SimpleStrategy' , 'replication_factor': '1'};
@@ -126,14 +50,6 @@ class CqlParserTest extends FlatSpec with Matchers {
     )
 
   }
-
-  def parseQuery(statement: String) =
-    CqlParser
-      .parseQuery(statement) match {
-        case CqlParser.Success(res: SelectStatement, _) => res
-        case CqlParser.Success(res, _)                  => fail(s"$res is not SelectStatement")
-        case CqlParser.Failure(msg, next)               => fail(s"Parse Failure: $msg, line = ${next.pos.line}, column = ${next.pos.column}")
-      }
 
   def parseSchema(statement: String) =
     CqlParser.parseSchema(statement) match {
